Allow validation of multiple attributes at once, with built in support for uniqueness checking of multiple columns
This commit allows you to support validating multiple attributes at
once:

  validates_each([:column1, :column2]) do |obj, attributes, values|
    # attributes = [:column1, :column2]
    # values = [obj.send(:column1), obj.send(:column2)]
  end

Support was added to validates_uniqueness_of to work on multiple
values at once:

  validates_uniqueness_of([:column1, :column2])

This is quite different from the following code:

  validates_uniqueness_of(:column1, :column2)

Which makes sure the each value is unique to its column, instead of
the combination of values being unique.

This will give a validation error message if an entry already exists
in the database that has the same value for column1 and column2.  It
works for any number of columns.

validates_uniqueness_of now supports an :allow_nil option that will
skip checking if the values of all columns are nil.  Previously, it
automatically skipped the validation if the value was blank.

There was also a slight memory reduction by reusing the default proc
if :if is not specified.