Don't have column accessor methods override plugin instance methods
This commit changes where the column accessor methods are defined.
Instead of defining them directly in the class, this defines them in
an anonymous module that is included in the class.  Because columns
accessors are generally defined when the class is created (before
the plugin has been added), the plugin instance methods module will
be included after the column accessor module, so it will take
precedence.  Because the column accessor methods don't override
existing instance methods, if a plugin has already been added, using
set_dataset after a plugin has been added (e.g. in a subclass) won't
create column accessors that override the plugin's instance methods.

This commit should also have a minor speedup in the general case, by
recognizing when a dataset is inherited and omitting some redundent
code in that case.

While here, simplify the code for Sequel::Model().