Disallow abuse of SQL function syntax for types (use :type=>:varchar, :size=>255 instead of :type=>:varchar[255])
This commit also refactors type_literal to return the entire literal
value for the type, instead of just the base type (so it includes
any length and/or UNSIGNED specifiers).  It also makes type_literal
private.

In addition, this fixes an issue with renaming or changing the type
of a varchar column on MySQL, if you wanted the size to be different
than 255.