Default to converting 2 digit years
Before, Sequel would treat 06/07/08 as being in the year 0008 instead
of 2008, at least when using the Date and DateTime classes.  This
commit changes that so that 2 digit years are treated as year + 1900
if year >= 69 or year + 2000 if year < 69.  To get back the old
behavior:

  Sequel.convert_two_digit_years = false

If you are using ruby 1.9, you should watch out, as the date parsing
differs from ruby 1.8:

  Date.parse('01/02/08', true).to_s # ruby 1.8 => "2008-01-02"
  Date.parse('01/02/08', true).to_s # ruby 1.9 => "2001-02-08"