Use INSERT ... RETURNING ... with PostgreSQL 8.2 and higher
This makes the PostgreSQL adapter use INSERT ... RETURNING ... if it
is connected to an 8.2.0 or later server.  This should provide a
performance boost, as it doesn't require extra queries being issued
to determine the value of the inserted primary key.

This also makes the Database object clear the cache for @primary_keys
and @primary_key_sequences when drop_table is called.  It adds a
Database#primary_key public method for determining the primary key
for a given table. If also adds a Dataset#server_version private
method to make things more testable.