Allow Models to use the RETURNING clause when inserting records on PostgreSQL
This commit makes model.save! check for an insert_select method on
the dataset.  If it exists, it calls that method and uses the result
as the values, instead of inserting and then reloading the values
from the database. This should be faster as well as less error prone.

It changes a recent commit to the PostgreSQL adapter so that
insert_sql does not use the RETURNING clause by default. You need to
call the insert_returning_sql method if you want the RETURNING
clause.  That is what Dataset#insert now does on PostgreSQL servers
8.2 and higher.  This caused some issues with prepared statements,
but this commit takes care of that as well.

This commit also contains a small speedup to Dataset#single_value,
as well as a missing spec for Database#raise_error.