Use string literals in AS clauses on SQLite
This commit adds a private Dataset method as_sql, and coverts the
various parts of Dataset that use the AS clause to use this method.
By default, as_sql uses an identifier for the alias, but the SQLite
adapter overrides it to use a string literal.

This commit also changes the remaining case where the AS keyword was
left out, when aliasing tables with a hash.  Now, the AS keyword is
used in that case.

This caused one part of a spec to be removed, the case of
@dataset.from(:a => :c[:d]).sql => "SELECT * FROM a c(d)". I'm not
sure that that is valid SQL anyway.