Make eager_graph respect associations' :order options (use :order_eager_graph=>false to disable)
This changes eager_graph so that it will add the columns specified in
each graphed association's :order option to the list of ordered
columns.  It qualifies the column with the table alias used for that
join.

This shouldn't cause problems unless:

1) You have an already qualified column in the :order option.

2) You have a column in the :order option that wasn't in the primary
   table (maybe because you were using the :eager_graph association
   option to include the table whose column you were using).

While remote, those problems can happen, and you can turn off the
including of the associations :order option by setting the
association's :order_eager_graph option to false.

This commit also fixes a bug in eager_graph when you were loading
an association with the same name as the base table in the FROM
clause.