Don't blindly qualify columns when eager graphing
Qualifying all entries in the :order option when eager_graphing is
not correct behavior.  For example, it's legimate to order by a
function, which shouldn't be qualified.  Also, already qualified
symbols should not be qualified again.  Change the qualifying to
only qualify unqualified symbols and SQL::Identifiers, either of
which may be wrapped in an SQL::OrderedExpression.  Do not qualify
anything else.