Don't default to localhost as the host in the PostgreSQL adapter
The PostgreSQL adapter used to default to 'localhost' if no host was
given.  Now, it doesn't specify a default, so the lower level software
(pg, ruby-postgres, or postgres-pr) will provide one.  The reason for
this is that PostgreSQL has always defaulted to using the unix socket
connection if available.  Sequel will now abide by this convention.

The PostgreSQL adapter still checks the contents of the :host option.
If it is blank (e.g. ''), it uses nil.  This is so it can continue to
work with the postgres-pr driver.  For the same reason, the default
port of 5432 is still specified in Sequel.