Don't modify a dataset's cached column information if calling #each with an option that modifies the columns
This was a significant though generally harmless bug.  It could be
easily triggered by the following code:

  dataset.count
  dataset.print

This is really a workaround for a design issue.  If each didn't
accept options, and had to clone the dataset and then run each to use
different options, this problem wouldn't have happened.  There are
places where that design would cause issues (e.g. #columns), but in
general it would cause fewer issues with the current design (which
has caused multiple bugs similar to this).

Anyway, we fix the problem by renaming #each to #_each and making it
a private method, then adding a new #each public method calling
_each in a begin/ensure block and resetting the @columns in the
ensure section if an opt was passed that modified the columns
returned.  This means that if you ask for the dataset columns inside
the call to each, you get the columns being returned in that each
call, but after each exits, @columns returns to its previous state.