Allow full text search on PostgreSQL to include rows where a search column is NULL
If performing a full text search on multiple columns, where one of
the columns is NULL but another column has all of the search words,
we want that row to show up.  To make that happen, we need to
COALESCE all of the search columns to turn NULLs into empty strings.

Also, we change full_text_search to use the simple language if no
language is specified, to be consistent with full_text_index.