Major performance enhancement when fetching rows using PostgreSQL
I did some profiling and found that fetching rows on PostgreSQL was
a significant bottleneck.  The previous code was decidely suboptimal,
this new code should be significantly faster.  Basically, the old
code was doing a hash lookup and calls to fname and ftype for every
column of every record, when those could be done in advance (since
they are the same for every record).  I've seen about a 3x
improvement in speed, and that's on fetches of only 150 records, it
should be greater on larger fetches.

Profiling also showed that Postgres.string_to_bool was a bottleneck.
Doing some research showed that postgresql always gives 't' for true,
and 'f' for false.  It shouldn't be called for NULL records, so the
conversation can be reduced to "== 't'", which is simple enough it
should not require a method, so I inlined it.