Refactor model hooks for increased performance
This change mainly stemmed from the fact that the call to
after_initialize in Model.initialize caused about a 20% performance
hit when fetching records, even if no after_initialize hook was
defined.  The main performance enhancing change in this commit is
that hook instance methods are defined to be empty until a hook class
method is called, which redefines the hook instance method of the
given type to actually call the hooks.

Refactor model hooks so they are inherited when the class is created,
at which point future hooks added to the superclass are not seen
in the subclass.  This makes sure that the @hooks class instance
variable is always available, except inside a block passed to
Class.new (see the spec changes in this commit for a work around).
It also simplifies Model.has_hooks?.  This breaks backwards
compatibility slightly, if you were adding hooks to a superclass
after a subclass had been created and expecting the hooks to appear
in the subclass.

This commit removes the Model.hooks and Model.all_hooks private
class methods.  It adds a Model.hook_blocks public class method that
yields all hook blocks for the given type of hook.  run_hooks is now
a private instance method instead of a private class method.