Lazy load model instance options such as raise_on_save_failure, for performance reasons
Before, 6 model instance options were loaded on initialize, which
slowed things down, as often it wasn't necessary to load them.
Now, don't load any instance option until it is necessary.

Add Module#class_attr_overridable to implement this.  It defines
an instance method which gets it default value from the class
method, but can be overridden using a standard attr_writer.

This commit breaks backwards compatibility by removing the
Model.lazy_load_schema setting.  That setting was necessary when
loading the schema was very slow and you were reloading code.  It
shouldn't be necessary now, as you can use DB.schema to get schema
for all tables at once, and even if you reload the model class, you
should be able to use the database's cached value (not to mention
that parsing the schema should be a fairly fast operation now).

This changes Model#initialize to not call a passed block if from_db
is true.  That should only be set by Model.load, which doesn't
use a block.  Also, Model#initialize now defaults to {} for the
values instead of nil, and passing nil now causes an error (this
is also a slight breakage of backwards compatibility).