Much better support for multiple schemas in PostgreSQL (Fixes #243)
Sequel already had decent support for selecting data in tables
outside the default schema, but it's support for schema manipulation
and introspection for tables outside the default schema was lacking.
This commit fixes that problem, though the approach could be taken
further (by inspecting the schema search path and operating on all
schemas in it).  I have no plans to add that functionality, but I'll
be happy to apply patches written by others who need it.

This commit does the following:

Makes the SQL generators schema-aware, though the default schema for
every database except PostgreSQL is nil and cannot be changed.

Explicitly uses the default schema when using PostgreSQL, which is
set to public by default and can be changed with default_schema=.
This breaks backwards compatibility for users who use a default
schema other than public.  If you use PostgreSQL and a schema
other than public, use DB.default_schema=, and any time you want to
access a table outside of that schema, make sure to qualify it with
a schema name.

Makes the primary key and sequence introspectors for PostgreSQL
schema aware.  The sequence introspectors now return a quoted,
qualified sequence.  This makes inserting into a table outside
a schema in the schema path return the primary key value.

Makes #table_exists? and #tables schema aware.  #tables now
accepts a block and yields a dataset if a block is given, instead
of returning an array of symbols.  #table_exists? uses this new
feature.

The @primary_keys and @primary_key_sequences instance variable
hashes are keyed on "schema"."table" instead of on table.