Support using any expression as the argument to Symbol#like
Before :a.like(:b) would do "a LIKE 'b'" instead of "a LIKE b". This
also didn't give the answer one would expect from code such as
:a.like(:c[:b]).

This breaks backwards compatibility, as before it would always treat
the passed object as a string.  However, I believe this is the more
correct way to handle things.