Add :primary_key option to one_to_many and many_to_one associations
This commit allows you to choose which key to use as the primary key
for one_to_many and many_to_one associations.  Generally, for
one_to_many associations, this is the primary key in the current
table, and for many_to_one associations, it is the primary key in the
associated table.  However, there may be cases where you want to
associate two models together, with a column in one model pointing to
a non-primary key column in another model.  This option allows for
that, and it's also supported when eager loading via eager or
eager_graph.

This commit also adds an eager_loader_key method to
AssociationReflection, removing some custom code from eager_load.
This should make adding custom associations easier.