Make Dataset#join_table take an option hash instead of a table_alias argument, add support for :implicit_qualifier option
This commit fixes a long standing bug when eager loading via
eager_graph.  If you are eagerly loading multiple associations for
the same model and you add any conditions with :graph_conditions or
:graph_only_conditions, all associations for that model other than
the first, would implicitly qualify the value conditions with the
last association table alias instead of the table alias for the
model.  This is hard to explain, so here's a demonstration:

  Album.many_to_one :band
  # Tracks with the same name as the album
  Album.one_to_many :same_name_tracks, :class=>:Track, \
    :graph_conditions=>{:name=>:name}
  Album.eager_graph(:band, :same_name_tracks).sql
  # => ... FROM albums
    LEFT OUTER JOIN bands AS band
      ON (band.id = albums.band_id)
    LEFT OUTER JOIN tracks AS same_name_tracks ON \
       ((same_name_tracks.album_id = albums.id) AND \
        (same_name_tracks.name = band.name))
                                 ^^^^

That's the mistake, as it should use albums.name instead of
band.name.  This commit fixes that, by adding an option to join_table
that allows explicitly specifying the implicit qualifier for the
values.

join_table already had a table_alias argument added.  In hindsight,
that was a mistake.  Change that argument to an option hash, keeping
backwards compatibility.  Add support for :implicit_qualifier option
to graph as well, which just passes it to join table.  Have
eager_graph use the :implicit_qualifer option, and no longer qualify
the keys manually in eager_graph.