Add #add_graph_aliases (select_more for graphs), and allow use of arbitrary expressions when graphing
Just like select_more adds columns to the already selected columns,
add_graph_aliases adds columns to the already graphed columns.

In addition, both set_graph_aliases and add_graph_aliases take an
optional third argument for each value in the hash, specifying an
arbitrary expression.  This expression is used in place of the
combination of the first and second arguments.  For example:

  ds.set_graph_aliases!(:a=>[:b, :c], :d=>[:e, :f, 42])
  # SELECT b.c AS a, 42 AS d FROM ...
  ds.first # => {:b=>{:c=>?}, :e=>{:f=>42}}