Optimize Date object creation in PostgreSQL adapter
This commit changes the PostgreSQL adapter to use Date.new instead of
Date.parse when it encounters a date field.  Date.new is 3-5 faster
according to my benchmarking.  Depending on the number of date fields
you have in your query, this can dramatically speed up fetches.  Note
that even with this commit, date fields still dramatically slow
fetches down, because the ruby Date class is slow to instantiate.

To implement this, add a apply_connection_settings method to
Postgres::Adapter, and have it run on initial connection as well as
any time the connection is reset.

You can turn off this optimization with
Sequel::Postgres.use_iso_date_format = false, which may be necessary
if you use a non-ISO date style and changing the date style to ISO
causes problems.