Transform AssociationReflection from a single class to a class hierarchy
This commit refactors association reflections from a single class to
a parent class (still called AssociationReflection), with a subclass
for each association type.  The AssociationReflection instance
methods previously had many checks for the association's type in
order to perform the correct behavior.  Type-specific behavior is one
of the main benefits of using classes.

This commit has some small breakage of backwards compatibility. The
ASSOCIATION_TYPES constant moves from instead AssociationReflection
to inside Associations, and becomes a hash with symbol keys and class
values, instead of an array of symbols.  The RECIPROCAL_ASSOCIATIONS
constant has been eliminated.

This commit does increase the amount of code, but it also makes the
code simpler.  In addition, it makes it much easier to add custom
association classes that are still supported by associate.  You
just add the type to ASSOCIATION_TYPES, the key being the
association name symbol, and the value being the custom association
class.