Fix PostgreSQL Date optimization so that it doesn't reject dates like 11/03/2008, bump version
It turns out the PostgreSQL optimization for dates went too far.
I should have only changed the output format specification, not
the input/output specification for year/month/day ordering.  It
turns out this breaks the standard American date format.  This is a
significant enough bug that it warrants a new release, so I'm also
bumping the version to 2.7.1.