Make Database#schema to handle implicit schemas on all databases and multiple identifier object types
This changes the key of the @schemas instance variable to be the
quoted schema/table value instead of the table symbol.  It keeps most
backwards compatibility by giving @schemas a default proc, so that
you can use any valid identifier object types as the key (valid
identifier object types are Symbol, String, Sequel::SQL::Identifier,
and Sequel::SQL::Qualified::Identifier).

This changes Database#table_exists? so that it no longer uses the
output of Database#tables.  The reason for this is that tables will
always give an array of Symbols, and table_exists? should be allowed
to take other identifier object types.  What it can do is check
the cached schema information for the database.  If the table is in
the cached schema information, the table exists.  If not, it does a
query.

Another change made is that the public schema will always be used if
you call DB.schema without a table argument on PostgreSQL, even if
you specify the :schema=>nil option. This is a small breakage to
backwards compatibility.

Another change made is that the MySQL adapter was updated so that
the schema methods don't give an error when using an SQL::Identifier.