Whether to upcase or quote identifiers can now be set separately
This changes Sequel by allowing the user to specify whether to upcase
identifiers.  Previously, upcasing was forced when quoting by default
(but not when using MySQL, PostgreSQL, or SQLite), and not used when
not quoting. This separates the quoting and upcasing functionality
and allows the user to choose one or the other, or both, or even turn
it on upcasing identifiers for database that didn't do so before.

Backwards compatibility is mostly taken care of.  The Sequel default
is still to upcase, but not when using the MySQL, PostgreSQL, or
SQLite adapters.  If you use a custom adapter and do not want to
upcase by default, add a private instance method to your database
subclass named upcase_identifiers_default and have it return false.

I also removed some useless RDoc for the Sequel.single_threaded
method, which hasn't really been valid for a long time.