Some small fixes for MySQL and for upcasing identifiers on JDBC
This changes Dataset#upcase_identifiers? to lazily load the setting.
Previously, it was handled in Database#initialize, but on JDBC (and
other adapters that have subadapters), the method that provides the
default hasn't been included in the class when Database#initialize
tried to use it, so this would end up with identifiers being upcased
when using MySQL, PostgreSQL, or SQLite on JDBC.

This fixes Database#alter_table for MySQL, for the drop_index case,
and fixes the MySQL schema parser to handle the fact that the table
name provided to schema_parse_table is now a string instead of a
symbol.