Make :schema__table.* work correctly
This shouldn't have been as difficult as it was, but the refactoring
done was for the best:

Move quote_schema_table and schema_and_table from Database to
Dataset, putting proxy methods in Database.

Make Database#default_schema a real attribute that can be used by
any database (but should remain nil for most).

Allow Database.new to take a :default_schema option, and document
it and other options that can be used.