Raise Errors if EXCEPT/INTERSECT is used when not supported
This adds a couple of modules to Sequel::Dataset that can be included
in adapters that don't support EXCEPT/INTERSECT.
UnsupportedIntersectExcept raises an error if EXCEPT/INTERSECT is
used, and UnsupportedIntersectExceptAll raises an error if
EXCEPT ALL/INTERSECT ALL is used (but EXCEPT/INTERSECT is allowed).
The Informix, MySQL, MSSQL, and Progress adapters use
UnsupportedIntersectExcept and the Oracle and SQLite adapters use
UnsupportedIntersectExceptAll.

Similar code was already in the MSSQL and Progress adapters, this
commit removes that code.  Also, it removes some commented out code
from the Informix adapter.