Add stored procedure support for the MySQL and JDBC adapters (Fixes #252)
This commit adds support for database stored procedures, with an API
similar to Sequel's prepared statement support, and implemented
internally in a similar way.  While it is directly callable on the
Database object (via #call_sproc), that use is discouraged.  Instead
it should be used at the dataset level with the following API:

  DB[:table].call_sproc(:select, :mysp, 'first param', 'second param')
  # or
  sp = DB[:table].prepare_sproc(:select, :mysp)
  sp.call('first param', 'second param')
  sp.call('third param', 'fourth param')

The only adapters with support for this are MySQL and JDBC (if using
a database that supports it).  Other databases don't even expose this
API.  Adding support to other databases should be fairly easy, though
I have no plans to at present.

The stored procedure implementation is similar to the prepared
statement implementation at the Dataset level.  #call_sproc and
and returning a clone of the dataset,and overriding Dataset#execute
and related functions to add options that are used by
Database#execute to send the request to the Dataset#call_sproc
method.

While working on stored procedure support, it became necessary to fix
the MySQL adapter to make it handle multiple results, since MySQL
stored procedures require that.  This also fixed issues with using
multiple statements at once in the MySQL adapter.  Before, this would
cause a "commands out of sync" error message that wasn't easily
recoverable from.  The MySQL adapter now supports this, though the
JDBC adapter connecting to MySQL still barfs when you attempt this.

Additionally, fix the socket tests in the MySQL adapter to use the same
user, password, and database.