Have the connection pool remove disconnected connections when the adapter supports it
This adds a Sequel::DatabaseDisconnectError (subclass of
Sequel::DatabaseError), for signaling to the connection pool that the
connection was lost.  It changes the connection pool code to rescue
that exception and remove the offending connection from the
connection pool.

In order to implement this, a disconnection_proc had to be added to
the connection pool.  Sequel uses a generic one that calls
Database#disconnect_connection.  disconnection_proc is called both
when connections are removed because the connection was lost and
when Database#disconnect is called.  Database#disconnect now calls
@pool.disconnect, which now uses the disconnection_proc if no
block is provided.  All adapters have been modified to remove
Database#disconnect and define Database#disconnect_connection, which
was fairly easy since all defined #disconnect methods just called
@pool.disconnect with a block that disconnected each connection.

The only adapter that currently supports this is PostgreSQL.  The
PostgreSQL adapter used to silently attempt to reconnect, which
might have caused the same SQL to be used twice. I'm not sure it
could have happened, but I'm not sure it couldn't have happened
either.  Now, if the database connection is lost, it raises
DatabaseDisconnectError, the connection pool removes the connection
from the pool, and raises the error to the application.  If the
application wants to continue, it can always retry.

While mucking in the connection pool, I found a bug where the
wrong key could be used when new connections were created.  This
wasn't a huge issue in most cases, but it could have caused as
many as twice the number of max_connections connections to be
created.