Correctly literalize BigDecimal NaN and (+-)Infinity values (Fixes #256)
This surrounds NaN and Infinity BigDecimal values in quotes when
literalizing them, which fixes errors on insert, since without quotes
they are generally treated as column references.  Database support
for these values is fairly limited, though.

This modifies the SQLite adapter to handle all three special values
correctly, and changes it to return BigDecimal whenever the column
type is numeric, decimal, or money.  This is backwards incompatible,
but makes SQLite support more consistant with other databases.

PostgreSQL works with NaN values, but errors if you try to give it
(+-)Infinity values unless the backing column is a floating point
column and not a numeric column.

As is typical for MySQL, it converts all three special values to 0.

No testing was done on the other adapters, but it is unlikely this
breaks something that worked before.