Default to logging only WARNING level messages when connecting to PostgreSQL
The PostgreSQL default for client_min_messages is NOTICE, but this
creates a lot of unnecessary messages (2 messages per SERIAL PRIMARY
KEY in a migration, for example).  Sequel now defaults to using
WARNING.  This can be changed via:

  Sequel::Postgres.client_min_messages = nil # Use Server Default
  Sequel::Postgres.client_min_messages = :notice # Use NOTICE level

This setting can only be used with the native PostgreSQL adapter.

This commit also adds RDoc documentation of the metaprogrammed
module level accessors used by the Postgres adapter to the Postgres
module.