Much better support for Dataset#{union,except,intersect}, allowing chaining and respecting order
This commit changes the internal dataset options uses to store
compound operations such as UNION, EXCEPT, and INTERSECT.  Before,
each compound operation was under its own key (2 keys really, since
it used to be :except and :except_all), and the value of the key only
stored a single dataset value, so chaining calls lost the previous
value. Also, there was a static order of compound operations, such
that some databases always had INTERSECT before UNION and some
databases always had UNION before INTERSECT.  This was a bad way
to do things that only handled simple cases well.

This commit changes that to use a single :compounds key that
stores an array of arrays of 3 items, compound type, dataset,
and an all flag.  When it comes time literalize the query,
the compound operations are applied in the order they were
called on the dataset.  Any order and number of compound operations
should work, as long as the database supports it.

By default, Sequel will use a subselect if using a compound
operation with a dataset that already has a compound operation on it.
That should work on most databases.  For databases that don't support
EXCEPT or INTERSECT, a very simple solution is used, as the order
of UNIONs shouldn't matter (they possibly matter if UNION and UNION
ALL or mixed, but we'll cross that bridge if we come to it).
PostgreSQL uses parantheses instead of a subselect to enforce
ordering of operations.

In addition to the spec changes, integration tests were also added
for this bugfix/feature.  They pass on SQLite, MySQL, and PostgreSQL,
though the MySQL tests are simple as they only test UNION.  Look
at the integration tests to see what is now possible.

This commit modifies the Informix, MSSQL, Oracle, and Progress
adapters without testing them, so if you run one of those databases,
be sure to test Sequel and make sure things still work.