Add define_hook method
This method allows to define own hook method for model. It will define 2
methods, ClassMethods#your_hook and InstanceMethods#your_hook. With
ClassMethods#your_hook you are able to pass block that should be called
when InstanceMethods#your_hook is invoked.

If InstanceMethods#your_hook returns false method that is invoking your
hook, method should be stoped. It's of course only internal standard, and
if you really don't like this behaviour you are not forced to use this.
But still it's highly recommended.