Add ability for Database#create_table to take options, support specifying MySQL engine, charset, and collate per table
This commit changes the Database#create_table and
Schema::Generator.new to take an options hash.  create_table
previously supported passing a generator as a second option, that is
now unofficially deprecated.  Now you should use the :generator
option.  I'm guessing the only code that passed a generator directly
was Model.create_table.  Some slight breakage of backwards
compatibility, as you are no longer allowed to pass nil as the second
option to create_table.  create_table! operates in much the same way.

Currently, the only adapter that uses the options is the MySQL
adapter.  It respects the :engine, :charset, and :collate options.
You can also specify defaults via:

  Sequel::MySQL.default_engine = 'InnoDB'
  Sequel::MySQL.default_charset = 'utf8'
  Sequel::MySQL.default_collate = 'utf8'

The defaults will be used if the options are not provided.  If a
default engine is set, you can specify :engine=>nil to not use it
(same goes for charset and collate).